Remove marketing blog code Separating the training planner app and the marketing site into separate repos.
