+++ README.md
@@ -1,10 +1,8 @@
-# Fit For Life Trainer &middot; [![License: WTFPL](https://img.shields.io/github/license/james-langridge/fit-for-life-trainer.svg)](http://www.wtfpl.net/about/) ![This repository is actively maintained](https://img.shields.io/maintenance/yes/2023)
+# A client management and scheduling app for personal trainers made with Next.js &middot; [![License: WTFPL](https://img.shields.io/github/license/james-langridge/fit-for-life-trainer.svg)](http://www.wtfpl.net/about/) ![This repository is actively maintained](https://img.shields.io/maintenance/yes/2023)
 
-Full stack web app for a personal trainer made with Typescript, Next.js v13, React v18, Tailwind CSS, Contentful (headless CMS) integration, PostgreSQL, and Prisma.
+Full stack web app for a personal trainer made with Next.js 13 (beta), React 18, Tailwind CSS, PostgreSQL, and Prisma.
 
-Site owner has complete flexibility to create and add content to new pages via Contentful, which are then statically generated (Static Site Generation), and revalidated on-demand when updated (Incremental Static Regeneration).
-
-Features include calendar app behind login. Admin can create, update, and delete training sessions in each client's calendar.  Clients can log in to follow the assigned training sessions on mobile or larger screens.  Calendar is optimised for mobile with infinite scroll.
+Admin can create, update, and delete training sessions in each users' calendar, with updates synced.  Users can log in to follow the assigned training sessions on mobile or larger screens.  Calendar is optimised for mobile with infinite scroll.
 
 ## Personal Trainer views
 ### Calendar
