--- components/contentful/index.ts
@@ -1,12 +1,0 @@
-export * from './CtfContactPage'
-export * from './CtaForm'
-export * from './CtfCtaSimple'
-export * from './CtfHeroBackgroundImage'
-export * from './CtfHeroImageHeading'
-export * from './CtfHeroSideImage'
-export * from './CtfImage'
-export * from './CtfProgramAd'
-export * from './CtfProgramAdBanner'
-export * from './CtfRichText'
-export * from './CtfTestimonial'
-export * from './CtfVideoHero'
