+++ middleware.ts
@@ -2,7 +2,7 @@ import {NextRequest, NextResponse} from 'next/server'
 import {jwtVerify} from 'jose'
 import {RequestCookies} from 'next/dist/compiled/@edge-runtime/cookies'
 
-// Unauthenticated users are redirected to /login from all (training-app) group routes in app/(training-app)/layout.tsx
+// Unauthenticated users are redirected to /login from all (training-app) group routes in app/(training-app)/layout-old.tsx
 
 const adminRoutes = ['/clients', '/register']
 
@@ -13,9 +13,9 @@ export default async function middleware(req: NextRequest) {
     const JWTPayload = await getJWTPayload(req.cookies)
 
     // This avoids redirecting unauthenticated users first to /training-planner
-    // then to /login via app/(training-app)/layout.tsx
+    // then to /login via app/(training-app)/layout-old.tsx
     if (!JWTPayload) {
-      req.nextUrl.pathname = '/login'
+      req.nextUrl.pathname = '/'
 
       return NextResponse.redirect(req.nextUrl)
     }
